Move mapping to the shape class
