+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/annotation/core/shapes/Rectangle.kt
@@ -1,6 +1,10 @@
 package com.github.barteksc.pdfviewer.annotation.core.shapes
 
 import android.graphics.PointF
+import com.github.salomonbrys.kotson.jsonNull
+import com.github.salomonbrys.kotson.jsonObject
+import com.github.salomonbrys.kotson.toJsonArray
+import com.google.gson.Gson
 import com.google.gson.JsonDeserializationContext
 import com.google.gson.JsonDeserializer
 import com.google.gson.JsonElement
@@ -15,15 +19,48 @@ data class Rectangle(
     override val points: List<PointF> = emptyList(),
     val edges: List<Edge> = emptyList(),
     val relations: Relations? = null,
-) : Shape(type, points)
-
-fun List<PointF>.generateRectangleEdges(): List<Edge> {
-    val edgeTopHorizontal = Edge(this[0], this[1])
-    val edgeRightVertical = Edge(this[1], this[2])
-    val edgeBottomHorizontal = Edge(this[2], this[3])
-    val edgeLeftVertical = Edge(this[3], this[0])
-    return listOf(edgeTopHorizontal, edgeRightVertical, edgeBottomHorizontal, edgeLeftVertical)
+) : Shape(type, points){
+    companion object{
+        fun generateRectangleEdges(points:List<PointF>): List<Edge> {
+            val edgeTopHorizontal = Edge(points[0], points[1])
+            val edgeRightVertical = Edge(points[1], points[2])
+            val edgeBottomHorizontal = Edge(points[2], points[3])
+            val edgeLeftVertical = Edge(points[3], points[0])
+            return listOf(edgeTopHorizontal, edgeRightVertical, edgeBottomHorizontal, edgeLeftVertical)
+        }
+    }
 }
+
+fun toJson(rectangle: Rectangle, gson: Gson) = jsonObject(
+    "type" to rectangle.type,
+    "points" to rectangle.points.map { point ->
+        jsonObject(
+            "x" to point.x.toString(),
+            "y" to point.y.toString(),
+            "z" to jsonNull
+        )
+    }.toJsonArray(),
+    "edges" to rectangle.edges.map {
+        jsonObject(
+            "start" to jsonObject(
+                "x" to it.start.x.toString(),
+                "y" to it.start.y.toString(),
+                "z" to jsonNull
+            ),
+            "end" to jsonObject(
+                "x" to it.end.x.toString(),
+                "y" to it.end.y.toString(),
+                "z" to jsonNull
+            ),
+            "value" to "0.0",
+            "unit" to "",
+            "distance" to jsonNull,
+            "name" to "",
+            "colorCode" to ""
+        )
+    }.toJsonArray(),
+    "relations" to gson.toJsonTree(rectangle.relations)
+)
 
 data class Edge(val start: PointF, val end: PointF)
